clean & flatten rewrite script
