--- scripts/replace-in-file/helpers/error-handler.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * Error handler
- */
-module.exports = function errorHandler(error, exitCode = 1) {
-  console.error(error);
-  process.exit(exitCode);
-};
