--- scripts/replace-in-file/helpers/success-handler.js
@@ -1,18 +1,0 @@
-'use strict';
-
-/**
- * Success handler
- */
-module.exports = function successHandler(results, verbose) {
-  const changed = results.filter(result => result.hasChanged);
-  const numChanges = changed.length;
-  if (numChanges > 0) {
-    console.log(`${numChanges} file(s) were changed`);
-    if (verbose) {
-      changed.forEach(result => console.log('-', result.file));
-    }
-  }
-  else {
-    console.log('No files were changed');
-  }
-};
