--- scripts/replace-in-file/version
@@ -1,1 +1,0 @@
-6.2.0
